backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.0.0
transformers==4.30.2
torch==2.0.1